The-ChatBot/requirements.txt
ipykernel
openai>=1.0.0
python-dotenv
yfinance
pandas
matplotlib